+++ game.h
@@ -262,15 +262,10 @@ bool has_any_player_food_collision(const game& g);
 ///Places a projectile in front of the player
 void put_projectile_in_front_of_player(std::vector<projectile>& projectiles, const player& p);
 
-int get_nth_food_timer(const game &g, const int &n);
-
 bool is_nth_food_eaten(const game &g, const int &n);
 
-coordinate get_nth_food_position(const game& g, const int& food_id);
 // Place a food item at a random position
 void place_nth_food_randomly(game &g, const int &n);
-
-coordinate get_nth_shelter_position(const game &g, const int &n);
 
 void test_game();
